[test] Fix wrong tooltip text detected by SWTBot test
The test TabBarTest fails randomly. The tooltip of the "Activate
Layouting Mode" button is sometimes setted with "&Zoom". This tooltip
corresponds to
org.eclipse.gmf.runtime.diagram.ui.l10n.DiagramUIMessages.ZoomActionMenu_ZoomLabel
that is used in ZoomExtensionContributionFactory.
The code in ZoomExtensionContributionFactory is not reliable. Indeed, we
set the tooltip of the last item of the tabbar and sometimes, this is
not the right one.
We now use the "index" parameter.

Change-Id: I672ef8db3ca03987f72f903160d23e307d7d4dbb
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>