[448398] Fix the LazyCrossReferencer behavior after a container change
A model element loses its LazyCrossReferencer adapter after changing the
container in case where the REMOVE/ADD notifications are handled in
reverse order. If the new container is already set and also has the
adapter then we add again the adapter.

Sometimes the model element contents and cross references are modified
before the REMOVE notification, in this case the inverse cross
referencer map is not properly cleaned. So we also remove target key
from the map in order to force future computation.

Bug: 448398
Change-Id: I439d93778cb66a857f374b6a143d8060aa37b620
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>