[445603] Fix possible deadlock in Modeling Projects
Now that the session saving requires an exclusive lock on the TED,
removing many semantic resources could lead to a deadlock when both the
session and the DefaultModelingProjectResourceListener try to react to
the removal.

DefaultModelingProjectResourceListener has no real reason to react to
resources removal: it was probably added to be symetrical with the
automatic attachment of new semantic resources but is redundant with
what the session already does (see DASI.removeResourceFromSession
triggered when statusesChanged() is called by ResourceSetSync), and in
some cases like here can lead to conflicts.

Bug: 445603
Change-Id: I8232597630bcfa848095a2fd9cc9b788325deb85
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>