[450227] Allow the creation of new class instead of LazyCrossReferencer
The anonymous class in
"DAnalysisSessionImpl.getSemanticCrossReferencer()" is replaced by the
new class "SessionLazyCrossReferencer". 

The method
"createSemanticCrossReferencer()" is added to instantiate the
"SessionLazyCrossReferencer" and can be overridden to allow the creation
of a specific ECrossReferenceAdapter 

The JUnit test "SiriusControlAndCrossReferenceTest" is modified to allow
the use of semantic cross referencer without internal member class.

This commit is a cherry-peak of fix of bug 449045.

Bug: 450227
Change-Id: Iafa3e0b59e6fd3892319b7c6155403f6b102ca36
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>