[445603] Return the result IStatus from DAnalysisSessionImpl.doSave()
Return the result of the save operation that may contain an exception
internally launched. The result can be used by the caller to analyze
errors during the save.

Indeed, exceptions that may be thrown during the save are no longer
propagated to the caller. So this fix allows sub-classes to have access
to the IStatus and react to it if they want to.

Bug: 445603
Change-Id: I858214108cf368397fc8c12697954de9150cbf41
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>