[445179] Fix original Draw2D bendpoint copy
The Draw2D bendpoints original locations were saved by copying a list.
Therefore, the list elements could be changed during the feedback. This
lead to a wrong restoration of the edge if a reconnection failed
(reconnected to a non candidate target). Note that this issue was
visible only if the GMF coordinates were broken, because Draw2D has to
compute coordinates.

Bug: 445179
Change-Id: I5226e7298bf812a906f0760f514ebbdb16192e44
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>