[475380] Use CanonicalDBorderItemLocator to fix a more precise location
Use CanonicalDBorderItemLocator to fix a more precise location and so a
more precise delta is stored in the request.
The CanonicalDBorderItemLocator has been improved to allow to set the
parent border bounds. Indeed, in our case, the parent bounds is not
already set to the corresponding bounds after move so we must set it to
allow the CanonicalDBorderItemLocator to consider it.

Bug:475380
Change-Id: I69e2048ee31fb0e7eea11ca45e75af8647dcb896
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>