[475380] Also arrange edges during the arrange of linked border nodes
Currently, there is no specific layout for edge during an arrange of
linked border nodes (with action Arrange All or with specific action
Arrange Linked Border Nodes). The result is strange.
This commit removes all bendpoints (except the first and the last) of
edges connected to a moved border node. The first and the last point are
set to the anchor location.

Bug: 475380
Change-Id: Iab5436a42103b3a711bfe1b751ea70c9fbe50249
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>