[454016] Use a ValidateEditSupport for TrackingModificationTrigger
Use a ValidateEditSupport for TrackingModificationTrigger to be sure to
be called once per transaction commited and not be called on rollback.
Use it with a delegate to manage case a previously existing
ValidateEditSupport.
In addition fix the issue of undo which doesn't update
Resource.isModified field.

Conflicts:
	plugins/org.eclipse.sirius.doc/doc/Release Notes.html
	plugins/org.eclipse.sirius.doc/doc/Release Notes.textile

Cloned from bug: 452653
Bug: 454016
Change-Id: I789bd34a9cff26947fa3a6c3cc5e380ddf8c0386
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>
(cherry picked from commit 5ca028139011f2c30380af9c58158e8821bc26c5)