[454003] tool variables are not displayed in autocompletion
variables are hard coded for the feature
ToolPackage.Literals.ABSTRACT_TOOL_DESCRIPTION__PRECONDITION in the
context of a Diagram.
This hard-coding is a temporary solution. The good way could be to put
that metadata on the diagram.ecore metamodeler as additional feature
annotations avoiding a specific code.

Note that the displayed order of the variables is not consistent between
connection start and complete Precondition. A specification should be
created to define the annotation display order according for example to
where it is in the metamodel hierarchy.

Cloned from bug: 428757
Bug 454003
Change-Id: I92aa16fdac3f5e18749e717456d66e12c37f7810
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>
(cherry picked from commit c70070c5931aa02004649cc246330d3ce3c5541a)