[445371] Split the SWTBot suite definition in two halves
The split point was chosen so that the two halves should be roughly of
half the duration of the complete suite. This first step does not change
anything to the actual content of the suite, but prepares for the
creation of two separate suites that can be run in parallel on a
CI-server with enough resources, thus (hopefuly) reducing the latency of
a complete test suite execution (the SWTbot suite is the one that
determines the latency).

Bug: 445371
Change-Id: I566d69159ccefefd5149d92db9033808649cd74f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>