[428752] Switch auto-completion to PROPOSAL_REPLACE mode (2/2 src-gen)
this commit is designed to backporting
this commit contains only generated source code

Fix auto-completion when case is not respected.
The content proposal adapter is switched to
ContentProposalAdapter.PROPOSAL_REPLACE This allows to replace the full
content of the expression allowing to change bad case, written by user,
by the good one proposed in autocompletion.

Bug 428752
Change-Id: I9d69a29c7772034a0cc3bcb95dc50fc14feed7cc
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>