[441642] Validate the diagram root element DDiagram
The validation is now applied to the diagram root element DDiagram.

Update INFO/WARNING/ERROR constraints registered in
org.eclipse.sirius.diagram/plugin.xml to add the target DDiagram.

Disable the validation decorator in SiriusValidationDecoratorProvider
for the DDiagram to avoid graphical conflict with other elements.

Update the DiagramValidationTest to check this new validation.

Bug: 441642
Change-Id: Ic50c682ebde7083ede25a30632cef94a611c9685
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>