[cleanup] Set session to null after SaveSessionJob execution
After a memory analysis, we have noticed that several AirDResourceImpl
are retained by DAnalysisSessionImpl, that themselves are retained by
SaveSessionJob. The job can indeed, be retained by ProgressManager.
The session is now reset to null after the job execution. Even if Job is
supposed to be reusable (see Javadoc of Job), the SaveSessionJob do
nothing if it is reused.

Change-Id: Ic9971c486e80d54ba185c69895bf926453d1f692
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>