[454001] Revert "Fix SemanticBasedDecorator domainClass property section"
This reverts commit e94f0e9f0272808dae3b6f090a5a18af28c8030f which uses
the TextContentProposalProvider.bindCompletionProcessor() to setup
completion in the domainClass field, but assumes that method has the new
behavior it has on master after commit
b4fd7ce859be1327d7569499a22436fdb1930b3e (as part of bug 428752), which
itself was *not* backported in v2.0.x.

This was a rather minor fix anyway, which was backported only because it
seemed innocuous (and by itself it is, the problem is its non-obvious
dependency towards another change which was not backported), so we'll
simply remove it from the scope of 2.0.2.

Bug: 454001
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>