[455589] Avoid NPE if an exception is thrown during savingPolicy.save()
Since bug 445603, there is a potential NPE if an exception is thrown
during the getSavingPolicy().save(). Indeed, the KO status of the
RunnableWithResult is overriden by OK_Status by the
getTransactionalEditingDomain().runExclusive(). So we can not use the
status of the RunnableWithResult after its execution. We use the result
instead: if null the runnable is considered as KO.

A test has been added to check the behavior when the savingPolicy fails.

Bug: 455589
Change-Id: I9115f54a87167251afb9eb622af4ce6a91dd59c3
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>