[456994] Switch to 'Warning' level to log a LockedInstanceException
The 'Error' level was used to log the message when catching a
LockedInstanceException exception. Switch it to the 'Warning' level. 

Add a "warnings" multimap on "SiriusTestCase" and
"AbstractSiriusSwtBotGefTestCase" to allow the recording of warning
logs. Some existing automated tests need to look at warnings to find
logs containing the "LockedInstanceException" exception.

Bug: 456994
Change-Id: Ib73d3a2b16106a615571c87ccf83a19c61b02569
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>