[454902] Add the script used for Gerrit verification
This is the script executed by the Gerrit verification job. It handles
the various combinations of branches, target platform, and test suite to
execute.

The Gerrit job itself fetches the latest version of this from the master
branch and then executes it, so the script on master must be able to
handle all the branches.

Bug: 454902
Change-Id: I57650c4611429e62f24b7347797432e50f430e37
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>