[424422] Fix underline and strike through styles on Linux
The underline and strike through font styles are not visible on Linux
because by default the line width is equal to 0.
In order to fix this for all the platforms and not to be dependent
anymore from the default value, when the underline or strike through
options are selected we set the line width to 1.

Bug: 424422
Change-Id: Ie908d2c0b8e2ade77eb16b9a89ad8ffcb3f3f945
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>