[457678] Minimize event ends computation during layout
Compute event ends only once and not three times.

We did not choose to introduce two new constructors to indicate the
allEventEnds at creation because the chosen way allows other subclassers
to
provide a way to compute differently the getAllEventEnds and to be sure
they will be called/computed during execute.

Bug: 457678
Change-Id: Ibd6cc121067ad1abfb1c4f65084d1b48273f60a3
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>