[457481] Check equality before storing a computed StyleDescription
Change the storage logic so that computed style description are reused
among style instances.
Also remove the listener responsible for cleaning up the "orphan" to
replace it by a dedicated method which can be called lower in the stack
and call it as part of the refresh.

The EMF modeled map
org.eclipse.sirius.diagram.ComputedStyleDescriptionRegistry.getCache()
Which was previously indexing the computed style descriptions
( EMap<DiagramElementMapping, EMap<EObject, EMap<EObject, EMap<EObject,
StyleDescription>>>> )
is no more used but not removed from the .ecore model yet. This will be
done in a distinct commit.

With EcoreTools and a class diagram having 500 EClasses, the number of
StyleDescription stored in the diagram itself went down from 9000 to 2/3
depending on the activated layers. In that case the refresh time was
also reduced by almost 40%.

Bug: 457481
Change-Id: Ia3b057470fe96f8d4ddf91c218dc198231498e84
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>