[424422] Update viewpoint metamodel to support underline and strike through in label format
Update in viewpoint metamodel the FontFormat enum to add underline and
strike through as possible formatting styles and remove the normal
value.
Update in viewpoint metamodel the BasicLabelStyleDescription labelFormat
attribute to a multi-valued attribute to be able to define a list of
font format for an element defining a label.
Update in table metamodel to remove the normal default value.
Update in tree metamodel the ForegroundStyleDescription labelFormat
attribute to a multi-valued attribute.
Update the specification editor properties view to support the new
styles.

This commit only contains MM changes so there are compilation problems.
They are expected.

Bug: 424422
Change-Id: Ie4c532f1696d0c596a96f87bd54ceced823ee655
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>