[457851] Version bumps based on semantic versioning
This commit bumps package versions depending on the API changes made
since the baseline.

The baseline used here is a Sirius 2.0.3 with the feature :
org.eclipse.sirius.runtime_2.0.3.201501210814

I used the prototype developped here: https://github.com/cbrun/baseliner
but as it rely on BND I'm sure it is doing a fairly good analysis.

pom.xml are modified only when the bundle version needs an upgrade.

Bug: 457851
Change-Id: I096ce586a47485c20e2ee27022026948c4e4388a
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>