[457851] Set target-provisionned API-baseline on projects
Introduce projects configurations which are being picked up by the
automatic OSGi baseliner. These configurations are refering to a target
platform definition (in the sirius.targets project) called 'baseline'
which defines the scope of the bundles to use as a baseline.

These settings have no impact if the baselining tooling is not
installed.

Bug: 457851
Change-Id: If981bd30cc69af4c1a7f73b277b950e079936ccd
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>