[444101] Remove the DTableElement adapters
Before this commit, DTableElements registered adapters on their semantic
target to be notified and refreshed if automatic refresh was
deactivated.
Now, the DTableElements are refreshed in pre-commit with the
RefreshEditorsPrecommitListener. This ModelChangeTrigger uses the new
RefreshImpactedElementsCommand and
DialectServices2.refreshImpactedElements() method. This method is not
currently added to the API DialectServices because it can probably
change for the implementation by other dialects (see for example the
gerrit [1] for bug 443518).

[1] https://git.eclipse.org/r/#/c/43165

Bug: 444101
Change-Id: I69e8797a0ded14c8a41b1c0b42e2affbc1f4ddd6
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>