[427247] Remove DNodeList.lineWidth attribute
It was never set and used only in DiagramListEditPartOperation (before
merge in DiagramContainerEditPartOperation). The lineWidth is the border
size and it is handled by the Style/StyleDescription
borderSize/borderSizeComputationExpression.

Furthermore, the default value was 1, and the code was trying to detect
the 0 value case to call setOutline(false) whereas for all kind of
DDiagramElement supporting the borderSize, we can find the following
code: 
if (boderSize == 0) 
  borderSize =1;
Removing this attribute and the corresponding code allow to remove dead
code and later questions on never used mecanism.

Bug: 427247
Change-Id: I06ae6fe69050a9f6f00d12d83cca26f0483d2f4f
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>