[459635] Do not create background figure twice
When the user use the set style to workspace image action on a
list/container, if the previous style was not an image, the
getBackgroundFigure() method returned null and the method
createBackgroundFigure() was called. Then the new background figure was
added to the figure returned by getFigure() which is not the good
container for the newly created WorkspaceImageFigure. This aim to create
a (SVG)WorkspaceImageFigure several times and to have a dummy figure
never used, refreshed, handled in the figure hierarchy.

The reinitFigure() method does all the needed job to init the background
figure and place it where it is expected.  

Bug: 459635
Change-Id: I06767f31412660b1b69975a22d202f96def94d5a
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>