[452962] Have Saver registration done in a single place and correctly
- Have Saver registration/unregistration for both
ResourceSetListener and Lifecycle in constructor/dispose().
- Add SaverTest to check that in case of rollback a save in postcommit
is
disarmed to avoid a save for the next executed command.
- SaverTest is minimalist and can be run in standalone, it's for that
SessionResourcesSynchronizer/DAnalysisSelectorService/DAnalysisSessionImpl
and some others classes are changed.
- To avoid "IllegalStateException: Cannot activate read/write
transaction in read-only transaction context" when calling the
SavingPolicy in postcommit and this SavingPolicy execute an EMF Command,
have Saver call the SavingPolicy in
TransactionalEditingDomainListener.transactionClosed().
Have WorkspaceBackEnd cleaned correctly on session close.

Bug: 452962
Change-Id: Ibca4ebdb0a85eaf1ab9c0583b49a6f8630174cc4
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>