[460816] Dispose graphical listeners sooner during close of diagram
Currently, the dispose of "graphical listeners"* are made in the
DDiagramEditorImpl.dispose(). But this method is finally called in
asyncExec by
org.eclipse.gmf.runtime.diagram.ui.resources.editor.parts.DiagramDocumentEditor.close(boolean).
This can be problematic because these listeners can be called with
inconsistent state during the close.

It is possible to dispose these graphical listeners directly in
DDiagramEditorImpl.dispose(), before the call to super.close(boolean),
to avoid this kind of problem.

"Graphical listeners": All the listeners that update the diagram
accordingly to notifications

Bug: 460816
Cherry-picked-from: 460432
Change-Id: Ib89780d37810ff888be8a7f3d9212a4bded65bf9
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>