[460000] Update tree item property sections only once per transaction
This fix improve performances when editing a tree item which is
contained in a large tree, for example, by using a representation tree
of 10k elements, this fix allows us to reduce the editing time of an
item for 33 seconds to 0.5 second.

This fix allows to update the tree items property sections only once per
transaction on post commit event instead for each item, which can be
very costly for large trees.

Bug: 460000
Change-Id: I7347e4ddb621f3a339065ae96bdd840691807c1c
Signed-off-by: Belqassim Djafer <belqassim.djafer@obeo.fr>