[442761] Fix arrange actions enablement after permission modification
After changing permissions to forbid diagram modification, arrange
actions were still available until the diagram gets the focus again that
triggers actions refresh.

DiagramActionContributionItem is now used to refresh the enablement of
the encapsulated arrange actions when it is updated.

Use a CopyOnWriteArrayList to store authority listeners in order to
avoid copying the list before iterating on it. Add notifications in
ReadOnlyPermissionAuthority to properly update listeners after locking
or unlocking some elements.

Bug: 442761
Change-Id: If7c209ad36bfd3d017f71a90bc4847b0c7a539b6
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>