[test] Dispose "Loading models" shell at tearDown
Sometimes this shell is not correctly disposed and has side effect on
following tests.
An error is log in this case. Indeed, the corresponding test should
probably be fixed.
If other shells also appear problematic, it will be possible to complete
the list SHELL_TO_CLOSE_AT_TEAR_DOWN_TEXTS.

Change-Id: Ifa83b6ece835c218cb52e494fbc155abeb9817bb
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>