[475821] Do not hide selected tree item in DTree editor
DTree editor do not have columns as expected by the
DTableEraseItemListener. This mechanism is only needed on DTables editor
to improve the selection feedback of the selected DCell and its Dline.

This commit also correct the active column computation and remove
duplicated code between table and tree.

Move the active column mecanism in DTable dialect only as the DTree
dialect does not have columns. This allows to remove duplicated and non
working code from DTree dialect.

This also allow to correct a expand/collapse bug in DTree editor on Mac
OS X, where user had to type twice on LEFT/RIGHT to expand or collapse
the tree items.

Bug: 475821
Change-Id: If55dc30d4b5bc2d8c806cfa2cdf13bfbf81548d0
Cherry-picked-from: 460810
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>
Signed-off-by: Esteban Dugueperoux <esteban.dugueperoux@obeo.fr>