[460810] Do not hide selected tree item in DTree editor
DTree editor do not have columns as expected by the
DTableEraseItemListener. This mechanism is only needed on DTables editor
to improve the selection feedback of the selected DCell and its Dline.

This commit also correct the active column computation and remove
duplicated code between table and tree. 

Move the active column mecanism in DTable dialect only as the DTree
dialect does not have columns. This allows to remove duplicated and non
working code from DTree dialect. 

This also allow to correct a expand/collapse bug in DTree editor on Mac
OS X, where user had to type twice on LEFT/RIGHT to expand or collapse
the tree items.

Bug: 460810
Change-Id: I7e6b3973ce82e7657a59e1167dcaf93db391c3d6
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>