[442491] Have DTreeItem expanding/collapsing done in a Job using monitor
Expanding a DTreeItem might be long if it has many direct children to
refresh, because this action refresh the content of the DTreeItem. For
example, having a DTreeItem with 10k children, the expand action of the
root element lasts 3 seconds.
This is why launching a progress monitor is suitable for the user to
have progression.

Bug: 442491
Change-Id: I052eee3d324696283e21428c1211a2ca4ca45d9f
Signed-off-by: Belqassim Djafer <belqassim.djafer@obeo.fr>