[427017] Avoid using the AirDCrossReferenceAdapter
AirDCrossReferenceAdapter is a "content adapter" which installs itself
on all elements of the aird resource. It is an extension of the GMF
CrossReferencerAdapter.

However considering the way we use GMF, we do not need this adapter at
all. To detect when GMF install it (as a collateral damage of the use of
a command, policy, API, ...), some checks have been added in the
abstract JUnit and SWTBot test cases. Thanks to this, our calls to
ViewUtil.destroy(View) have been detected, these calls are not needed
with Sirius: a simple EcoreUtil.remove() is sufficient in the detected
case, then the DanglingRefRemovalTrigger will do the same cleanup than
ViewUtil but without the installation of the GMF CrossReferencerAdapter.

Remove also GMF note attachments when the option "Remove/Hide not when
the annotated element is removed/hide" is disabled. 

Bug: 427017
Change-Id: I7fe516b6ed2473cf222c345492a7814421e67b9f
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>