[460947] Utilities to support invocation of Java from the workspace
Introduce utility classes and APIs abstracting ClassLoading capabilities
used by query interpreters.

An Interpreter registers callbacks on an instance of
JavaExtensionsManager.
These callbacks will be triggered when a class has been loaded,
unloaded, notfound or gone missing.
The JavaExtensionsManager maintains its state in sync with the current
list of imports, the scope (projects, plugins) and the workspace
notifications.
By using the same mechanisms, the JavaExtensionsManager computes and
maintain the list of EPackages declarations which are visibles in the
current classpath.

The JavaExtensionsManager is designed in a way which makes an instance
shareable among different interpreters if one wants to avoid the cost of
search/loading several times for the same class.

It relies on a ClassLoading implementation (BundleClassLoading by
default) which searches for classes in the current Bundles. This
implementation is overriden by WorkspaceClassLoading when the
org.eclipse.sirius.editor plugin is
installed.
This implementation also searches in workspace classes and uses the PDE
to collect dependencies.

Bug: 460947
Change-Id: I5ce7454f1d7bf168b5a408612ce73afd79154119
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>