[460947] Make ServiceInterpreter use JavaExtensionsManager
This brings support for Java services loaded from the workspace in the
'service:' interpreter.
Code completion is working.
I had to add logic to support de-registering of previously registered
services as in the context of the workspace a service might be updated.

The Interpreter only register a callback to the JavaExtensionsManager
instance even if it is owned by the Interpreter right now, but this is
done this way so that later on we can easily keep only one
JavaExtensionsManager which would be shared by several interpreters.

Bug: 460947
Change-Id: I8dd83fed107630a4db4afb298c0900fe044ede5a
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>