[427017] Avoid using the AirDCrossReferenceAdapter
AirDCrossReferenceAdapter is a "content adapter" which installs itself
on all elements of the aird resource. It is an extension of the GMF
CrossReferencerAdapter. However considering the way we use GMF, we do
not need this adapter at all.

Use ViewDeleteCommand instead of the GMF DeleteCommand in order to call
"EcoreUtil.remove(EObject)" instead of "ViewUtil.destroy(View)". The
DanglingRefRemovalTrigger will do the cleanup without the installation
of the GMF CrossReferenceAdapter.

Bug: 427017
Change-Id: I126ddc4aa60813a0f2c532d5be602e18f3a8b228
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>