[460351] Correct NPE during navigation from problem marker
When the user clicks on an error marker, this trigger the goto marker
mecanism. 

If the "marked" editor is not opened, the "error editor" is opened
during editor initalization but replaced during goto marker navigation. 
This is done because Eclipse tries to open a Sirius diagram editor using
a basic FileEditorInput referencing only the file, with no information
on the concrete representation to open.
+ Addition of a test that opens a diagram using a marker

Bug: 460351
Change-Id: Ie5f8fefd3728f4878325cd7a454ba163db4a4b32
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>