[427017] Remove AirdCrossReferenceAdapter
This was a "content adapter" which installs itself on all elements of
the aird resource. It was made to customize GMF's own
org.eclipse.gmf.runtime.emf.core.util.CrossReferenceAdapter.

However considering the way we use GMF, we do not need this adapter at
all.

Since the aird resource is not traversed anymore by the cross reference
adapter, fix some parts of the semantic resource detection during the
session opening.

Bug: 427017
Change-Id: Idf18aad2974c4a5575aed837d6dc52d602210eca
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>