[452962] Move creation of IWorkspaceRunnable
The creation of a IWorkspaceRunnable in the SavingPolicy is not the
right place. Indeed, this can lead to deadlock problem (see [1] for
detail). The creation of a IWorkspaceRunnable has been moved to the
Saver itself. By doing that, the lock on the Workspace is asked before
the lock on Saver.isSaving (same order as in case of a session.save()
called from a WorkspaceModifyOperation).

[1] https://bugs.eclipse.org/bugs/show_bug.cgi?id=452962#c21

Bug: 452962
Change-Id: I88e54d8c7b535f16350daca7a61c0c7316ce43c4
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>