[432133] Add test XSDSemanticResourceTests
Add tests about a semantic resource that conforms to a MM generated with
an XSD.

Ensure that the semantic model is detected after the addition of the
semantic resource or an Aird resource referencing the semantic resource.

Ensure that the session is dirty after a modification on the semantic
model.

Bug: 432133
Change-Id: I1769b4db1c9eece58b01a8288333168251044bc4
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>