[462213] Properly remove the semantic root from aird model tag
A map is used to maintain reference between the root of the semantic
resource which is set as models in aird AND its resource which contains
it.
This allows to properly remove the root from the aird model tag when its
resource is empty due to unload because of an external deletion for
instance.

Bug: 462213
Change-Id: Ifb47d4baf2e6be64f99000400a17ed3862f80475
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>