[460947] Register EPackages from the classpath in AQL
Use the JavaExtensionsManager infrastructure to compute the list of
EPackages which are available in the current classpath instead of
initializing AQL by registering all the EPackages which are installed in
the current platform.

This avoid an important initialization cost where AQL would trigger the
loading of every registered EPackage (and their own classpath)
This also prevent the side effect that the static analysis of AQL has to
analyse all those EPackages and then be slower than expected.

Bug: 460947
Change-Id: I7aa823ad3cbad10332e939ed7a6a71539be0fb02
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>