[443518] Use DialectServices2.refreshImpactedElements() to refresh DDE
The goal of this commit is to use the new implementation, as for table
and tree in bug 444101, to refresh diagram elements when the refresh
mode is manual.

* The method IDiagramElementEditPart.getEAdapterSemanticElements() has
been removed (and all its implementations).
* The DiagramDialectServices now implements DialectServices2, and method
refreshImpactedElements(DRepresentation, Collection<Notification>,
IProgressMonitor).
* The EObjectQuery has now a new contructor with a
ECrossReferenceAdapter parameter. This can be usefull if there is many
call to inverse references methods for the same Session. This avoids to
compute ECrossReferenceAdapter each time. It also proposes a new method
getInverseReferences(Set<EReference>).

Bug: 443518
Bug: 444101
Change-Id: Ic6306cd8e3f51d98507e1cdacc67e5179f9a0094
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>