[461498] Register Sirius's packages during expression validation
Use the IInterpretedExpressionQuery infrastructure to add the dialect
EPackages in the list of EPackages an interpreter should use to validate
the expressions.

For each Dialect the 'Representation' EPackage is added as it is
generaly used even in variable types. Each 'Description' package is also
added under the rational that it is not so uncommon to have expressions
using the representationElement->Mapping reference to check names for
instance.

Bug: 461498
Change-Id: I2109ddb07f54d806baa083b83350119584c8e826
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>