[474334] Systematically declare Dialect Metamodels as available
Make sure all the dialects metamodels are always available to
expressions.

This avoids the Acceleo compilation cache to be polluted with versions
of the service modules with compilation errors when they are compiled in
some context A, and then picked up in another context B, where they
would be fine but are not recompiled in the new context.

Bug: 474334
Change-Id: I6d697637d489e23854b60b41d74e0f77c2d02c1c
See-also: 461498
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>