[458822] Remove deep walk inside semantic model in the delete command
Clean DeletionCommandBuilder to avoid iterating inside semantic
elements. The permission authority now is responsible to iterate through
the contents of a semantic element if needed to allow or forbid the
deletion. 

IPermissionAuthority.canDeleteInstance(EObject) also checks that the
container can be edited, so calls to
IPermissionAuthority.canEditInstance(EObject) are unnecessary.

Do not add all the content of a semantic element in the list of element
to destroy. The content will be automatically removed with the
containing element if the action is triggered.

Bug: 458822
Change-Id: Id27800daae3e58e051351245fb37c38f894fdc03
Signed-off-by: Mickael LANOE <mickael.lanoe@obeo.fr>