[427247] Remove DValidable
It contains a single validate() EOperation and is inherited by
DDiagram and DDiagramElement, but:
* DDiagram.validate() throws UnsupportedOperationException("Should no
  more be called")
* The only implementation for a sub-type od DDiagramElement which is
  actually called somewhere is for DEdge (in DEdgeSpec.validate(),
  called from EdgeMappingHelper.updateEdge).

It's not worth defining an abstraction in the metamodel for a single
use case in the code base, The type should be removed from the
metamodel, along with all the unused validate() implementations (and
associated code which would become dead after that). The
DEdgeSpec.validate() implementation, if if is indeed useful, should be
moved as a helper method in EdgeMappingHelper, as it is the only place
that calls it.

Bug: 427247
Change-Id: I63b17ef0703903026d9cf37a9dea480a00b29fa7
Signed-off-by: Laurent Fasani <laurent.fasani@obeo.fr>