[463485] Add a snap to all shapes mode
* Several edit parts have been updated to use SiriusSnapToHelperUtil
instead of SnapToHelperUtil to choose SiriusSnapToGeometry instead of
SnapToGeometryEx
* SiriusSnapToGeometry extends SnapToGeometryEx to allow to snap to all
visible shapes and not only brothers ones. It uses the snapToAll mode
from extended data of request (set by the below trackers).
* Update tracker NoCopyDragEditPartsTrackerEx to activate the mode
snapToAllShapes when shortcut is pressed.
* Update tracker SiriusResizeTracker to activate the mode
snapToAllShapes when shortcut is pressed. A specific request is used for
this tracker as we can not use the extended meta-data, as for
NoCopyDragEditPartsTrackerEx. Indeed, the meta-data is cleaned before
calling SiriusSnapToGeometry.
* Add automatic tests with zoom, scroll in diagram, scroll in container
with all kind of nodes (container, node, border nodes).
* GraphicalHelper.getAbsoluteBoundsIn100Percent has been improved
because of a lack of precision detected during tests. All similar
methods, ie using translateToAbsolute(), have been also improved.

Bug: 463485
Change-Id: Icaabf7d7d763c9cf2a283e3cab76f1ec0aa0956e
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>