[463297] Correct DNodeListSpec.getElements() and reduce code duplication
Update DNodeListSpec.getElements() to be consistent with the
DNodeContainerSpec.getElements() method: when the methods were written,
it was not possible to specify border nodes for lists. This is supported
now, so the DNodeList.getElements has to return the border nodes and the
list elements.

Move all border nodes refresh from spec classes to a common helper. 

Bug: 463297
Change-Id: Id1cc8af9af6bb2427da52e42859538e2980653a4