[436359] Handle label styles for DNodeList and regions
When a label style is detected, do not draw the compartment border.

AbstractDiagramElementContainerEditPart.createNodePlate() is no more
abstract and now has a common implementation for lists and containers
which is the previous container implementations.

This allow to use the bevelled corner label style for Containers,
RegionContainers, Regions and DNodeList. 

FreeForm container and List container mappings with no label styles keep
their previous behaviors.

Bug: 436359
Change-Id: I40ed6669b76e88ddd09d251ced5101b9b1ff890b
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>