[463456] Add multiline support in property sections of interpreted expressions.
Modify all property sections of interpreted expression to have a push
button next to the text field that open a dialog containing a text area
with content proposal.
Also add an abstract property section
(AbstractTextWithButtonPropertySection) which is very similar to
AbstractTextPropertySection but with a button next to the text field and
an abstract method to creation the listener on the button.

Bug: 463456
Change-Id: I514ab9f0af6a711b8eea449d79d86ff2f7fa0f01
Signed-off-by: Benjamin Grouhan <benjamin.grouhan@obeo.fr>