[460706] Add ImagePath as a new DataType in the metamodel
The DataType of properties for image paths were only recognized as
String objects, but this type was not precise enough to differentiate
them from other String properties. Changing the DataType of these
properties to ImagePath will help to recognize them (needed in the next
commit).
Also modify the constraint for validating image paths to make it in a
single class.
Moreover, in QuickOutlineControl, the attributes to search in were
indicated directly via the datatypes, but, with this change in the
metamodel, it would not be correct anymore. So, to fix this (and avoid
further fixes when modifying another DataTypes whose correspond to
String objects), the condition now tests the instance type of the
attributes.

Bug: 460706
Change-Id: I0b2a6acf0db41da1527a411d3ee337335d5a8722
Signed-off-by: Benjamin Grouhan <benjamin.grouhan@obeo.fr>