[427799] Simplify DASI.toString()
The previous implementation of DASI.toString() was complex and
potentially costly. In particular it relied on calling allAnalysis(),
which triggers the loading of all the referenced aird resources if they
are not already loaded. Given the many contexts in which toString() can
be called (e.g. just selecting an instance in the debugger), such an
important side-effect is a bad idea.

Bug: 427799
Change-Id: Idfa1a8de50f2cddcf1009447641ca2989a3b8076
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>