[427799] Do not check for session validity in SessionManager.add()
The current criterion for session validity is rather weak, and at the
same time it relies on the fact that all the representation resources
are loaded (and will load them if they are not already). Anyway, it
should probably be the responsibility of the session itself not to try
to register itself in the SessionManager if it is invalid.

Bug: 427799
Change-Id: I8ef8703b2334e02c62e37af919951682df4c0036
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>