[427799] Simplify DASI.getID()
The only place where this was called is in SessionManagerImpl.add(), and
then only to check against Session.INVALID_SESSION. That call has been
removed (see previous commit), so there is no reason anymore to have
this relatively complex and costly code (costly because it requires that
all the aird resources are loaded). Keep the getID() method for now as
it is API, but simply use the main session resource's URI as identifier.

Bug: 427799
Change-Id: Id8bef06f1547bf9c188e7b95b1aebae69e33ae99
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>